Add finetunning all checkpoints

+++ finetunning/mtet_allcheckpoint.py
@@ -144,7 +144,7 @@ checkpoints = [int(x.replace('.index', '').split('-')[-1]) for x in tf.io.gfile.
 checkpoints.sort()
 
 NUM_WORKER = 3
-CHUNK_LENGTH = len(checkpoints) / NUM_WORKER
+CHUNK_LENGTH = int(len(checkpoints) / NUM_WORKER)
 checkpoints = [checkpoints[i:i + CHUNK_LENGTH] for i in range(0, len(checkpoints), CHUNK_LENGTH)]
 CHUNK = args.chunk
